♻️ extract settings to own crate
