+++ Cargo.toml
@@ -3,7 +3,7 @@ resolver = "2"
 
 default-members = ["client"]
 
-members = ["client", "services/index-git-repositories"]
+members = ["client", "services/index-git-repositories", "settings"]
 
 [workspace.package]
 version = "1.1.0"
