+++ client/src/plugin/utils.rs
@@ -164,38 +164,11 @@ pub trait Plugin {
     }
 }
 
-pub fn config_directory() -> anyhow::Result<String> {
-    let home_directory = std::env::var("HOME")?;
-    let config_in_home = format!("{home_directory}/.config");
-    Ok(std::env::var("XDG_CONFIG_HOME").unwrap_or(config_in_home))
-}
-
-pub fn centerpiece_default_config_path() -> anyhow::Result<String> {
-    let config_directory = crate::plugin::utils::centerpiece_config_directory()?;
-    Ok(format!("{config_directory}/config.yml"))
-}
-
-pub fn centerpiece_config_directory() -> anyhow::Result<String> {
-    let config_directory = config_directory()?;
-    Ok(format!("{config_directory}/centerpiece"))
-}
-
-pub fn cache_directory() -> anyhow::Result<String> {
-    let home_directory = std::env::var("HOME")?;
-    let cache_in_home = format!("{home_directory}/.cache");
-    Ok(std::env::var("XDG_CACHE_HOME").unwrap_or(cache_in_home))
-}
-
-pub fn centerpiece_cache_directory() -> anyhow::Result<String> {
-    let cache_directory = cache_directory()?;
-    Ok(format!("{cache_directory}/centerpiece"))
-}
-
 pub fn read_index_file<T>(file_name: &str) -> anyhow::Result<T>
 where
     T: serde::de::DeserializeOwned,
 {
-    let cache_directory = centerpiece_cache_directory()?;
+    let cache_directory = settings::centerpiece_cache_directory()?;
     let index_file_path = format!("{cache_directory}/{file_name}");
 
     let index_file =
